ITimer interface cleanup
