+++ Midi/Sanford.Multimedia.Midi.csproj
@@ -253,10 +253,11 @@
     <Compile Include="Sanford.Collections\PriorityQueue.cs" />
     <Compile Include="Sanford.Collections\SkipList.cs" />
     <Compile Include="Sanford.Multimedia.Timers\ThreadTimer.cs" />
-    <Compile Include="Sanford.Multimedia.Timers\GenericTimer.cs" />
     <Compile Include="Sanford.Multimedia.Timers\ITimer.cs" />
     <Compile Include="Sanford.Multimedia.Timers\Time.cs" />
     <Compile Include="Sanford.Multimedia.Timers\Timer.cs" />
+    <Compile Include="Sanford.Multimedia.Timers\TimerFactory.cs" />
+    <Compile Include="Sanford.Multimedia.Timers\TimerQueue.cs" />
     <Compile Include="Sanford.Multimedia\Device.cs" />
     <Compile Include="Sanford.Multimedia\DeviceException.cs" />
     <Compile Include="Sanford.Multimedia\ErrorEventArgs.cs" />
